requirements.txt
# Web Framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# REST API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LLM Integration
openai>=1.3.0
google-generativeai>=0.3.0

# Data Processing & VCF Parsing
pandas>=2.0.0
#pysam>=0.21.0

# Visualization & Utils
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# API & HTTP
requests>=2.31.0

# Data Serialization
PyPDF2>=3.0.0
pyyaml>=6.0

# Database (optional for caching)
sqlalchemy>=2.0.0